convert CollectorsAdministration to typescript and using React functional component
